Addressed latest review comments
- Moved auto-refresh JS to Scripts/auto-refresh.js embedded resource
- Replaced [index..] range syntax with .Substring()
- Added using for Configuration namespace, removed inline qualifications
- Kept InvariantCulture on float.TryParse (RFC 7231 requires period decimal)
- Removed static from private methods in non-static classes
- Renamed static fields to s_ prefix (s_regexTimeout, s_sharedHttpClient)
- Replaced target-typed new() with explicit new Type()
- Fixed PackageReferences to use $(JellyfinNugetVersion) instead of $(JellyfinVersion)
- Added early return when embedded script resource is missing
